requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
mysqlclient==2.2.7
mysql-connector-python
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
Pillow==11.2.1
Flask-Mail==0.9.1
weasyprint
gunicorn==21.2.0
psycopg2-binary==2.9.9 